requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2==2.9.9
djangorestframework-simplejwt==5.2.2
black==23.10.1